requirements.txt
InquirerPy==0.3.4
python-dotenv==1.0.1
Requests==2.32.3
rich==13.9.4
